requirements.txt
fastapi==0.90.0
uvicorn==0.20.0
gunicorn==20.1.0

Jinja2==3.1.2

python-dotenv==0.21.1 # to load the configurations in `.env` file
email-validator==1.3.1 # used with pydantic, to get EmailStr
python-multipart==0.0.5 # manging streaming uploads and User Forms

# database connection / migration / ORM
alembic==1.9.3
SQLAlchemy==2.0.3
# psycopg2==2.9.5 # for windows / local installation
psycopg2-binary==2.9.5 # for linux

# password hashing
passlib==1.7.4
bcrypt==4.0.1

# for JWT token, alternative: python-jose
PyJWT==2.6.0
